+++ python/sglang/srt/models/openvla.py
@@ -3,25 +3,12 @@ from typing import Iterable, List, Optional, Tuple, Any, Dict, ClassVar, Union
 import numpy as np
 import torch
 from torch import nn
-from transformers import (
-    CLIPVisionConfig,
-    CLIPVisionModel,
-    LlavaConfig,
-    PretrainedConfig
-)
+from transformers import PretrainedConfig
 from vllm.config import CacheConfig
 from vllm.model_executor.layers.quantization.base_config import QuantizationConfig
-from vllm.model_executor.model_loader.weight_utils import default_weight_loader
 
-from sglang.srt.mm_utils import (
-    get_anyres_image_grid_shape,
-    unpad_image,
-    unpad_image_shape,
-)
-from sglang.srt.model_executor.forward_batch_info import ForwardMode, InputMetadata
+from sglang.srt.model_executor.forward_batch_info import InputMetadata
 from sglang.srt.models.llama2 import LlamaForCausalLM
-from sglang.srt.models.mistral import MistralForCausalLM
-from sglang.srt.models.qwen2 import Qwen2ForCausalLM
 from transformers.models.auto import CONFIG_MAPPING
 
 import logging
@@ -36,273 +23,33 @@ import torch
 import torch.nn as nn
 import transformers
 from timm.models.vision_transformer import LayerScale
-from transformers import AutoModelForCausalLM, PretrainedConfig, PreTrainedModel
-from transformers.modeling_outputs import ModelOutput, CausalLMOutputWithPast
+from transformers import PretrainedConfig, PreTrainedModel
+from transformers.modeling_outputs import ModelOutput
 
 import timm.data
 import torch
-import torchvision.transforms.functional as TVF
-from PIL import Image
-from torchvision.transforms import CenterCrop, Compose, Normalize, Resize, ToTensor
-from transformers import PreTrainedTokenizerBase
-from transformers.image_processing_utils import BatchFeature, ImageProcessingMixin
-from transformers.processing_utils import ProcessorMixin
-from transformers.tokenization_utils import PaddingStrategy, PreTokenizedInput, TextInput, TruncationStrategy
-from transformers.utils import TensorType
-
-import time
 
 TIMM_OVERRIDE_ACT_LAYER: Dict[str, List[Optional[str]]] = {
-    "clip-vit-l": ["quick_gelu"], "clip-vit-l-336px": ["quick_gelu"],
-    "dinov2-vit-l": [None], "in1k-vit-l": [None],
-    "siglip-vit-so400m": [None], "siglip-vit-so400m-384px": [None],
+    "clip-vit-l": ["quick_gelu"],
+    "clip-vit-l-336px": ["quick_gelu"],
+    "dinov2-vit-l": [None],
+    "in1k-vit-l": [None],
+    "siglip-vit-so400m": [None],
+    "siglip-vit-so400m-384px": [None],
     "dinoclip-vit-l-336px": [None, "quick_gelu"],
-    "dinosiglip-vit-so-224px": [None, None], "dinosiglip-vit-so-384px": [None, None]
+    "dinosiglip-vit-so-224px": [None, None],
+    "dinosiglip-vit-so-384px": [None, None],
 }
 
-LLM_BACKBONE_TO_HF_METACLASS = {"llama2-7b-pure": "llama",}
+LLM_BACKBONE_TO_HF_METACLASS = {
+    "llama2-7b-pure": "llama",
+}
 
 logger = logging.getLogger(__name__)
 
 # === PyTorch/HuggingFace Default IGNORE_INDEX (for CrossEntropyLoss labels)
 IGNORE_INDEX = -100
 
-# ========== Preprocessor ==========
-
-# === Image Processing ===
-def letterbox_pad_transform(image: Image.Image, padding_fill_value: Tuple[int, int, int]) -> Image.Image:
-    """Given a PIL.Image, pad to square by adding a symmetric border around the height/width."""
-    (w, h), max_wh = image.size, max(image.size)
-    horizontal_pad, vertical_pad = int((max_wh - w) / 2), int((max_wh - h) / 2)
-    padding = (horizontal_pad, vertical_pad, horizontal_pad, vertical_pad)
-
-    return TVF.pad(image, padding, fill=padding_fill_value, padding_mode="constant")
-
-
-class PrismaticImageProcessor(ImageProcessingMixin):
-    model_input_names: ClassVar[List[str]] = ["pixel_values"]
-
-    def __init__(
-        self,
-        use_fused_vision_backbone: bool = False,
-        image_resize_strategy: str = "letterbox",
-        input_sizes: Optional[List[Tuple[int, int, int]]] = None,
-        interpolations: Optional[List[str]] = None,
-        means: Optional[List[Tuple[float, float, float]]] = None,
-        stds: Optional[List[Tuple[float, float, float]]] = None,
-        **kwargs: str,
-    ) -> None:
-        """
-        Initialize a PrismaticImageProcessor as a wrapper around a torchvision transform; this transform will be
-        created by TIMM, and edited to follow our custom `image_resize_strategy` logic.
-        @param use_fused_vision_backbone: Boolean indicating single or fused (dual) vision backbone
-        @param image_resize_strategy: Prismatic image resize strategy in < resize-naive | resize-crop | letterbox >
-        @param input_size: [TIMM :: `data_cfg`] Input image size as tuple (channels, width, height)
-        @param interpolation: [TIMM :: `data_cfg`] Interpolation as string (default: "bicubic")
-        @param mean: [TIMM :: `data_cfg`] Normalization mean as float tuple (or two-tuple if `fused_backbone`)
-        @param std: [TIMM :: `data_cfg`] Normalization std as float tuple (or two-tuple if `fused_backbone`)
-        """
-        self.use_fused_vision_backbone = use_fused_vision_backbone
-        self.image_resize_strategy = image_resize_strategy
-
-        # Handle `None` default values
-        input_sizes = [(3, 224, 224)] if input_sizes is None else input_sizes
-        means = [(0.5, 0.5, 0.5)] if means is None else means
-        stds = [(0.5, 0.5, 0.5)] if stds is None else stds
-
-        # TIMM `data_cfg` Parameters
-        self.input_sizes, self.interpolations, self.means, self.stds = input_sizes, interpolations, means, stds
-
-        # Grab torchvision transforms via TIMM =>> need to parse for specific "functional" transform values!
-        self.tvf_resize_params, self.tvf_crop_params, self.tvf_normalize_params = [], [], []
-        self.tvf_do_letterbox, self.tvf_letterbox_fill = False, None
-
-        for idx in range(len(input_sizes)):
-            transform = timm.data.create_transform(
-                input_size=self.input_sizes[idx],
-                interpolation=self.interpolations[idx],
-                mean=self.means[idx],
-                std=self.stds[idx],
-                crop_pct=1.0,  # Set to 1.0 to ignore cropping (initial Resize sets `input_size`)
-                crop_mode="center",  # Default crop mode -- no-op when `crop_pct == 1.0`
-                is_training=False,  # No image augmentations when loading the transform!
-            )
-
-            # [Validation] Ensure appropriate transform structure, expected sizes
-            if not (
-                isinstance(transform, Compose)
-                and (len(transform.transforms) == 4)
-                and isinstance(transform.transforms[0], Resize)
-                and isinstance(transform.transforms[1], CenterCrop)
-                and isinstance(transform.transforms[2], ToTensor)
-                and isinstance(transform.transforms[3], Normalize)
-                and (transform.transforms[0].size == self.input_sizes[idx][-1])
-                and (transform.transforms[1].size == self.input_sizes[idx][-2:])
-            ):
-                raise ValueError(f"Unexpected TIMM image transformation structure/sizes: `{transform}`")
-
-            # HF Image Processors *must* be JSON-serializable; as such, cannot have torchvision. as an attribute.
-            #   => Instead, we're going to parse the transform and call "torchvision.transforms.functional" (`tvf`)
-            resize_t, crop_t, norm_t = transform.transforms[0], transform.transforms[1], transform.transforms[3]
-            self.tvf_resize_params.append(
-                {
-                    "size": resize_t.size,
-                    "interpolation": TVF.pil_modes_mapping[resize_t.interpolation],
-                    "max_size": None,
-                    "antialias": True,
-                }
-            )
-            self.tvf_crop_params.append({"output_size": crop_t.size})
-            self.tvf_normalize_params.append(
-                {
-                    "mean": norm_t.mean.float().numpy().tolist(),
-                    "std": norm_t.std.float().numpy().tolist(),
-                    "inplace": False,
-                }
-            )
-            self.tvf_do_letterbox, self.tvf_letterbox_fill = False, None
-
-            # Handle Prismatic `image_resize_strategy`
-            if self.image_resize_strategy == "resize-naive":
-                self.tvf_resize_params[idx]["size"] = (resize_t.size, resize_t.size)
-            elif self.image_resize_strategy == "letterbox":
-                self.tvf_do_letterbox, self.tvf_letterbox_fill = True, tuple([int(x * 255) for x in self.means[idx]])
-            elif self.image_resize_strategy == "resize-crop":
-                pass
-            else:
-                raise ValueError(f"Image resize strategy `{self.image_resize_strategy}` is not supported!")
-
-        # Dispatch **kwargs to super()
-        super().__init__(**kwargs)
-
-    def apply_transform(self, img: Image.Image) -> torch.Tensor:
-        """Apply `functional` variant of TIMM's Transform = Compose([Resize -> CenterCrop -> ToTensor -> Normalize])"""
-        if self.tvf_do_letterbox:
-            img = letterbox_pad_transform(img, self.tvf_letterbox_fill)
-
-        # [Contract] Fused Backbones expect "channel-stacked" inputs; we'll unpack on the model side!
-        imgs_t = []
-        for idx in range(len(self.input_sizes)):
-            img_idx = TVF.resize(img, **self.tvf_resize_params[idx])
-            img_idx = TVF.center_crop(img_idx, **self.tvf_crop_params[idx])
-            img_idx_t = TVF.to_tensor(img_idx)
-            img_idx_t = TVF.normalize(img_idx_t, **self.tvf_normalize_params[idx])
-            imgs_t.append(img_idx_t)
-
-        # [Contract] `imgs_t` is a list of Tensors of shape [3, input_size, input_size]; stack along dim = 0
-        img_t = torch.vstack(imgs_t)
-
-        return img_t
-
-    def preprocess(
-        self,
-        images: Union[Image.Image, List[Image.Image]],
-        return_tensors: Optional[Union[str, TensorType]] = None,
-        **_: str,
-    ) -> BatchFeature:
-        """
-        Preprocess an image (or batch of images); note that unlike the `transformers :: BaseImageProcessor` we
-        explicitly only handle PIL.Image.Image instances for simplicity.
-        @param images: A (batch of) PIL.Image.Image instance(s) to preprocess.
-        @param return_tensors: BatchFeature default Tensor format (e.g., "pt" for torch); if None, returns np.ndarray
-        @return: Instance of `transformers :: BatchFeature` with a single key "pixel_values"
-        """
-        if not isinstance(images, list):
-            images = [images]
-
-        # Apply `self.img_transform` to each image (will return list of torch.Tensors); stack into "batched" Tensor
-        pixel_values = torch.stack([self.apply_transform(img.convert("RGB")) for img in images])
-
-        # Return BatchFeature =>> note that for compatibility, constructor expects Dict[str, np.ndarray], so we convert
-        return BatchFeature(data={"pixel_values": pixel_values.float().numpy()}, tensor_type=return_tensors)
-
-    def __call__(self, images: Union[Image.Image, List[Image.Image]], **kwargs) -> BatchFeature:
-        return self.preprocess(images, **kwargs)
-
-
-# === PrismaticProcessor =>> Wraps both ImageProcessor and Tokenizer ===
-#   =>> https://github.com/huggingface/transformers/blob/main/src/transformers/models/llava/processing_llava.py
-class PrismaticProcessor(ProcessorMixin):
-    attributes: ClassVar[List[str]] = ["image_processor", "tokenizer"]
-    image_processor_class: str = "AutoImageProcessor"
-    tokenizer_class: str = "AutoTokenizer"
-
-    def __init__(
-        self,
-        image_processor: Optional[ImageProcessingMixin] = None,
-        tokenizer: Optional[PreTrainedTokenizerBase] = None,
-    ) -> None:
-        super().__init__(image_processor, tokenizer)
-
-    def __call__(
-        self,
-        text: Union[TextInput, PreTokenizedInput, List[TextInput], List[PreTokenizedInput]],
-        images: Union[Image.Image, List[Image.Image]],
-        padding: Union[bool, str, PaddingStrategy] = False,
-        truncation: Optional[Union[bool, str, TruncationStrategy]] = None,
-        max_length: Optional[int] = None,
-        return_tensors: Optional[Union[str, TensorType]] = TensorType.PYTORCH,
-    ) -> BatchFeature:
-        """
-        Preprocess a given (batch) of text/images for a Prismatic VLM; forwards text to the underlying LLM's tokenizer,
-        forwards images to PrismaticImageProcessor.
-        @param text: The (batch) of text to encode; must be a string or list of strings.
-        @param images: A (batch of) PIL.Image.Image instance(s) to preprocess.
-        @param padding: Sequence padding strategy (if multiple specified) in < True = "longest" | "max_length" | False >
-        @param truncation: Truncation strategy for the output sequences; requires `max_length` to be specified
-        @param max_length: Maximum length (in tokens) to truncate
-        @param return_tensors: Type of return tensors (usually "pt" or TensorType.PYTORCH)
-        @return: BatchFeature with keys for `input_ids`, `attention_mask` and `pixel_values`.
-        """
-        pixel_values = self.image_processor(images, return_tensors=return_tensors)["pixel_values"]
-        text_inputs = self.tokenizer(
-            text, return_tensors=return_tensors, padding=padding, truncation=truncation, max_length=max_length
-        )
-
-        # [Validate] Need same number of images and text inputs!
-        if pixel_values.shape[0] != text_inputs.input_ids.shape[0]:
-            raise ValueError("Batch is malformed; expected same number of images and text inputs!")
-
-        return BatchFeature(data={**text_inputs, "pixel_values": pixel_values})
-
-    # === Tokenizer Dispatch Utilities =>> check `PreTrainedTokenizerBase` for documentation ===
-    def batch_decode(
-        self,
-        sequences: Union[List[int], List[List[int]], torch.Tensor, Any],  # `Any` = np.ndarray | tf.Tensor
-        skip_special_tokens: bool = False,
-        clean_up_tokenization_spaces: Optional[bool] = None,
-        **kwargs: str,
-    ) -> List[str]:
-        return self.tokenizer.batch_decode(
-            sequences=sequences,
-            skip_special_tokens=skip_special_tokens,
-            clean_up_tokenization_spaces=clean_up_tokenization_spaces,
-            **kwargs,
-        )
-
-    def decode(
-        self,
-        token_ids: Union[int, List[int], torch.Tensor, Any],  # `Any` = np.ndarray | tf.Tensor
-        skip_special_tokens: bool = False,
-        clean_up_tokenization_spaces: Optional[bool] = None,
-        **kwargs: str,
-    ) -> str:
-        return self.tokenizer.decode(
-            token_ids=token_ids,
-            skip_special_tokens=skip_special_tokens,
-            clean_up_tokenization_spaces=clean_up_tokenization_spaces,
-            **kwargs,
-        )
-
-    @property
-    def model_input_names(self) -> List[str]:
-        tokenizer_input_names = self.tokenizer.model_input_names
-        image_processor_input_names = self.image_processor.model_input_names
-
-        return list(dict.fromkeys(tokenizer_input_names + image_processor_input_names))
-
-
-# ========== Model ==========
 
 # === Utility Functions for Monkey-Patching ===
 def unpack_tuple(fn: Callable[[Any], Tuple[Any]]) -> Callable[[Any], Any]:
@@ -312,16 +59,19 @@ def unpack_tuple(fn: Callable[[Any], Tuple[Any]]) -> Callable[[Any], Any]:
 
     return wrapper
 
+
 # HF Transformers overwrites parameters with names containing `gamma`; we're going to patch VisionBackbone.LayerScale.
 #   =>> TIMM :: https://github.com/huggingface/pytorch-image-models/blob/main/timm/models/vision_transformer.py#L109
 #   =>> Transformers :: https://github.com/huggingface/transformers/blob/main/src/transformers/modeling_utils.py#L3960
 def _ls_new_forward(self, x: torch.Tensor) -> torch.Tensor:
     return x.mul_(self.scale_factor) if self.inplace else x * self.scale_factor
 
+
 def ls_apply_patch(ls_module: LayerScale):
     ls_module.scale_factor = nn.Parameter(ls_module.gamma.clone())
     ls_module.forward = _ls_new_forward.__get__(ls_module, LayerScale)
     del ls_module.gamma
+
 
 # === Prismatic Vision Backbone (nn.Module) Definitions (w/ Fused Backbone Support) ===
 class PrismaticVisionBackbone(nn.Module):
@@ -338,7 +88,9 @@ class PrismaticVisionBackbone(nn.Module):
         # [Contract] Validate number of (fused) vision backbones, create "alpha" featurizer and Instantiate
         #   =>> Note :: Monkey-Patch the `forward()` function of the backbone to ensure FSDP-compatibility
         #               Hardcodes `get_intermediate_layers` to return the **SECOND-TO-LAST** layer patches!
-        assert len(timm_model_ids) <= 2, "Prismatic models only support up to 2 (fused) vision backbones!"
+        assert (
+            len(timm_model_ids) <= 2
+        ), "Prismatic models only support up to 2 (fused) vision backbones!"
         self.featurizer = timm.create_model(
             timm_model_ids[0],
             pretrained=False,
@@ -347,7 +99,10 @@ class PrismaticVisionBackbone(nn.Module):
             act_layer=timm_override_act_layers[0],
         )
         self.featurizer.forward = unpack_tuple(
-            partial(self.featurizer.get_intermediate_layers, n={len(self.featurizer.blocks) - 2})
+            partial(
+                self.featurizer.get_intermediate_layers,
+                n={len(self.featurizer.blocks) - 2},
+            )
         )
         self.embed_dim = self.featurizer.embed_dim
 
@@ -361,7 +116,10 @@ class PrismaticVisionBackbone(nn.Module):
                 act_layer=timm_override_act_layers[1],
             )
             self.fused_featurizer.forward = unpack_tuple(
-                partial(self.fused_featurizer.get_intermediate_layers, n={len(self.fused_featurizer.blocks) - 2})
+                partial(
+                    self.fused_featurizer.get_intermediate_layers,
+                    n={len(self.fused_featurizer.blocks) - 2},
+                )
             )
             self.embed_dim += self.fused_featurizer.embed_dim
 
@@ -379,7 +137,9 @@ class PrismaticVisionBackbone(nn.Module):
         """Run image (`pixel_values`) through featurizer; if channel-stacked, then dispatch and sequence stack."""
         print("===== pixel values =====")
         print(type(pixel_values))
-        pixel_values = torch.from_numpy(pixel_values).unsqueeze(0).to(torch.bfloat16).to('cuda')
+        pixel_values = (
+            torch.from_numpy(pixel_values).unsqueeze(0).to(torch.bfloat16).to("cuda")
+        )
         if not self.use_fused_vision_backbone:
             return self.featurizer(pixel_values)
 
@@ -389,8 +149,11 @@ class PrismaticVisionBackbone(nn.Module):
 
         return torch.cat([patches, patches_fused], dim=2)
 
+
 class PrismaticProjector(nn.Module):
-    def __init__(self, use_fused_vision_backbone: bool, vision_dim: int, llm_dim: int) -> None:
+    def __init__(
+        self, use_fused_vision_backbone: bool, vision_dim: int, llm_dim: int
+    ) -> None:
         super().__init__()
         self.use_fused_vision_backbone = use_fused_vision_backbone
         self.vision_dim, self.llm_dim = vision_dim, llm_dim
@@ -422,6 +185,7 @@ class PrismaticProjector(nn.Module):
 
         return projected_features
 
+
 # === Main HF Class Definitions ===
 @dataclass
 class PrismaticCausalLMOutputWithPast(ModelOutput):
@@ -437,12 +201,16 @@ class PrismaticCausalLMOutputWithPast(ModelOutput):
     # Additions for VLMs
     projector_features: Optional[torch.FloatTensor] = None
 
+
 class OpenVLAConfig(PretrainedConfig):
     model_type: str = "openvla"
     is_composition: bool = False
+
     def __init__(
         self,
-        norm_stats: Optional[Dict[str, Dict[str, Dict[str, Dict[str, List[float]]]]]] = None,
+        norm_stats: Optional[
+            Dict[str, Dict[str, Dict[str, Dict[str, List[float]]]]]
+        ] = None,
         n_action_bins: int = 256,
         vision_backbone_id: str = "siglip-vit-so400m",
         llm_backbone_id: str = "vicuna-v15-7b",
@@ -468,10 +236,16 @@ class OpenVLAConfig(PretrainedConfig):
         self.use_fused_vision_backbone = (
             use_fused_vision_backbone
             if use_fused_vision_backbone is not None
-            else any(self.vision_backbone_id.startswith(v) for v in ["dinoclip", "dinosiglip"])
+            else any(
+                self.vision_backbone_id.startswith(v)
+                for v in ["dinoclip", "dinosiglip"]
+            )
         )
 
-        self.timm_model_ids = ["vit_large_patch14_reg4_dinov2.lvd142m", "vit_so400m_patch14_siglip_224"]
+        self.timm_model_ids = [
+            "vit_large_patch14_reg4_dinov2.lvd142m",
+            "vit_so400m_patch14_siglip_224",
+        ]
         self.timm_override_act_layers = TIMM_OVERRIDE_ACT_LAYER[self.vision_backbone_id]
         self.image_sizes = [224, 224]
         self.image_resize_strategy = image_resize_strategy
@@ -482,7 +256,9 @@ class OpenVLAConfig(PretrainedConfig):
 
         # [IMPORTANT] HF Utilities actually look for a `text_config` field... we need to use that specific naming!
         self.text_config = (
-            CONFIG_MAPPING[LLM_BACKBONE_TO_HF_METACLASS[self.llm_backbone_id]](**text_config)
+            CONFIG_MAPPING[LLM_BACKBONE_TO_HF_METACLASS[self.llm_backbone_id]](
+                **text_config
+            )
             if text_config is not None
             else CONFIG_MAPPING[LLM_BACKBONE_TO_HF_METACLASS[self.llm_backbone_id]]()
         )
@@ -494,6 +270,7 @@ class OpenVLAConfig(PretrainedConfig):
         # Dispatch **kwargs to super() =>> note that `pad_token_id` collides, so we pass it in here as well...
         super().__init__(pad_token_id=pad_token_id, **kwargs)
 
+
 class OpenVLAForActionPrediction(PreTrainedModel):
     config_class: PretrainedConfig = OpenVLAConfig
     base_model_prefix: str = "model"
@@ -501,6 +278,7 @@ class OpenVLAForActionPrediction(PreTrainedModel):
     _no_split_modules: ClassVar[List[str]] = ["PrismaticProjector"]
     _skip_keys_device_placement: str = "past_key_values"
     _supports_flash_attn_2: bool = True
+
     def __init__(
         self,
         config: OpenVLAConfig,
@@ -518,7 +296,9 @@ class OpenVLAForActionPrediction(PreTrainedModel):
                 "TIMM Version must be >= 0.9.10 and < 1.0.0 (breaking); please raise a GitHub Issue "
                 "if you urgently need support for latest TIMM versions."
             )
-        if (transformers.__version__ != "4.40.1") or (tokenizers.__version__ != "0.19.1"):
+        if (transformers.__version__ != "4.40.1") or (
+            tokenizers.__version__ != "0.19.1"
+        ):
             logger.warning(
                 f"Expected `transformers==4.40.1` and `tokenizers==0.19.1` but got "
                 f"`transformers=={transformers.__version__}` and `tokenizers=={tokenizers.__version__}`; "
@@ -528,7 +308,10 @@ class OpenVLAForActionPrediction(PreTrainedModel):
 
         # Instantiate PrismaticVisionBackbone (w/ Potential Fused Backbone)
         self.vision_backbone = PrismaticVisionBackbone(
-            config.use_fused_vision_backbone, config.image_sizes, config.timm_model_ids, config.timm_override_act_layers
+            config.use_fused_vision_backbone,
+            config.image_sizes,
+            config.timm_model_ids,
+            config.timm_override_act_layers,
         )
 
         # Create Multimodal Projector
@@ -539,7 +322,9 @@ class OpenVLAForActionPrediction(PreTrainedModel):
         )
 
         # Instantiate LLM Backbone
-        self.language_model = LlamaForCausalLM(config.text_config, quant_config=quant_config)
+        self.language_model = LlamaForCausalLM(
+            config.text_config, quant_config=quant_config
+        )
         # self.language_model = AutoModelForCausalLM.from_config(
         #     config.text_config, attn_implementation=config._attn_implementation
         # )
@@ -558,7 +343,9 @@ class OpenVLAForActionPrediction(PreTrainedModel):
         self.bin_centers = (self.bins[:-1] + self.bins[1:]) / 2.0
 
         # Compute vocab size for de-tokenization -- revert added "multiple of"
-        self.vocab_size = self.config.text_config.vocab_size - self.config.pad_to_multiple_of
+        self.vocab_size = (
+            self.config.text_config.vocab_size - self.config.pad_to_multiple_of
+        )
 
         print("===== Using OpenVLAForActionPrediction =====")
 
@@ -618,11 +405,13 @@ class OpenVLAForActionPrediction(PreTrainedModel):
 
         if pixel_values is not None:
             pixel_values = pixel_values[0]
-        
+
         # === Handle Unimodal Forward ===
         if pixel_values is None:
             print("[Unimodal generation]")
-            assert input_ids is not None, "Missing `input_ids` in language-only forward!"
+            assert (
+                input_ids is not None
+            ), "Missing `input_ids` in language-only forward!"
             language_model_output = self.language_model(
                 input_ids=input_ids,
                 positions=positions,
@@ -645,12 +434,17 @@ class OpenVLAForActionPrediction(PreTrainedModel):
 
             # Build Multimodal Embeddings & Attention Mask =>> Prismatic defaults to inserting after <BOS> token (1:)
             multimodal_embeddings = torch.cat(
-                [input_embeddings[:, :1, :], projected_patch_embeddings, input_embeddings[:, 1:, :]], dim=1
+                [
+                    input_embeddings[:, :1, :],
+                    projected_patch_embeddings,
+                    input_embeddings[:, 1:, :],
+                ],
+                dim=1,
             )
 
             # Dispatch to Language Model
             multimodal_embeddings = multimodal_embeddings.squeeze(0)
-            positions = torch.arange(1, multimodal_embeddings.shape[0]+1)
+            positions = torch.arange(1, multimodal_embeddings.shape[0] + 1)
             input_metadata.seq_lens = torch.tensor([multimodal_embeddings.shape[0]])
             input_metadata.positions = positions
             print("input_metadata", input_metadata)
@@ -663,12 +457,15 @@ class OpenVLAForActionPrediction(PreTrainedModel):
                 input_metadata=input_metadata,
                 input_embeds=multimodal_embeddings,
             )
-        
+
         # language_model_output.next_token_logits = language_model_output.logits
         return language_model_output
-    
+
     def predict_action(
-        self, input_ids: Optional[torch.LongTensor] = None, unnorm_key: Optional[str] = None, **kwargs: str
+        self,
+        input_ids: Optional[torch.LongTensor] = None,
+        unnorm_key: Optional[str] = None,
+        **kwargs: str,
     ) -> np.ndarray:
         """Thin wrapper around super().generate() that decodes predicted actions and de-normalizes them."""
 
@@ -676,30 +473,46 @@ class OpenVLAForActionPrediction(PreTrainedModel):
         # (after "OUT:" or "ASSISTANT:"), insert it to match the inputs seen at training time
         if not torch.all(input_ids[:, -1] == 29871):
             input_ids = torch.cat(
-                (input_ids, torch.unsqueeze(torch.Tensor([29871]).long(), dim=0).to(input_ids.device)), dim=1
+                (
+                    input_ids,
+                    torch.unsqueeze(torch.Tensor([29871]).long(), dim=0).to(
+                        input_ids.device
+                    ),
+                ),
+                dim=1,
             )
         print(kwargs)
         # Run VLA inference
-        generated_ids = self.generate(input_ids, max_new_tokens=self.get_action_dim(unnorm_key), **kwargs)
+        generated_ids = self.generate(
+            input_ids, max_new_tokens=self.get_action_dim(unnorm_key), **kwargs
+        )
         # generated_ids = self.generate(input_ids, max_new_tokens=7, **kwargs)
 
         # Run VLA inference without KV-Cache
         # generated_ids = input_ids.clone()
         # self.past_key_values = None
-        # for _ in range(7):  
+        # for _ in range(7):
         #     output = self.generate(generated_ids, max_new_tokens=1, **kwargs)
         #     generated_ids = torch.cat((generated_ids, output[:, -1:]), dim=-1)
 
         # Extract predicted action tokens and translate into (normalized) continuous actions
-        predicted_action_token_ids = generated_ids[0, -self.get_action_dim(unnorm_key) :].cpu().numpy()
+        predicted_action_token_ids = (
+            generated_ids[0, -self.get_action_dim(unnorm_key) :].cpu().numpy()
+        )
         discretized_actions = self.vocab_size - predicted_action_token_ids
-        discretized_actions = np.clip(discretized_actions - 1, a_min=0, a_max=self.bin_centers.shape[0] - 1)
+        discretized_actions = np.clip(
+            discretized_actions - 1, a_min=0, a_max=self.bin_centers.shape[0] - 1
+        )
         normalized_actions = self.bin_centers[discretized_actions]
 
         # Unnormalize actions
         action_norm_stats = self.get_action_stats(unnorm_key)
-        mask = action_norm_stats.get("mask", np.ones_like(action_norm_stats["q01"], dtype=bool))
-        action_high, action_low = np.array(action_norm_stats["q99"]), np.array(action_norm_stats["q01"])
+        mask = action_norm_stats.get(
+            "mask", np.ones_like(action_norm_stats["q01"], dtype=bool)
+        )
+        action_high, action_low = np.array(action_norm_stats["q99"]), np.array(
+            action_norm_stats["q01"]
+        )
         actions = np.where(
             mask,
             0.5 * (normalized_actions + 1) * (action_high - action_low) + action_low,
@@ -709,7 +522,9 @@ class OpenVLAForActionPrediction(PreTrainedModel):
         return actions
 
     @staticmethod
-    def _check_unnorm_key(norm_stats: Dict[str, Dict[str, Any]], unnorm_key: Optional[str]) -> str:
+    def _check_unnorm_key(
+        norm_stats: Dict[str, Dict[str, Any]], unnorm_key: Optional[str]
+    ) -> str:
         if unnorm_key is None and len(norm_stats) != 1:
             raise ValueError(
                 f"Your model was trained on more than one dataset. "
@@ -718,7 +533,9 @@ class OpenVLAForActionPrediction(PreTrainedModel):
             )
 
         # If None, grab the (singular) dataset in `norm_stats` to use as `unnorm_key`
-        unnorm_key = unnorm_key if unnorm_key is not None else next(iter(norm_stats.keys()))
+        unnorm_key = (
+            unnorm_key if unnorm_key is not None else next(iter(norm_stats.keys()))
+        )
         if unnorm_key not in norm_stats:
             raise ValueError(
                 f"The `unnorm_key` you chose ({unnorm_key = }) is not in the available statistics. "
@@ -742,7 +559,9 @@ class OpenVLAForActionPrediction(PreTrainedModel):
         # print("===== Get Input Embeeding =====")
         # torch.save(self.language_model.get_input_embeddings(), "weights/embedding_layer.pt")
         if self.embeddings_layer == None:
-            self.embeddings_layer = torch.load("weights/embedding_layer.pt", weights_only=False).to('cuda')
+            self.embeddings_layer = torch.load(
+                "weights/embedding_layer.pt", weights_only=False
+            ).to("cuda")
         # print(torch.equal(embeddings_layer.to('cpu').weight, self.language_model.get_input_embeddings().to('cpu').weight))
         return self.embeddings_layer
         # return self.language_model.get_input_embeddings()
@@ -766,7 +585,9 @@ class OpenVLAForActionPrediction(PreTrainedModel):
         if ((input_ids is not None) and (input_ids.shape[0] > 1)) or (
             (inputs_embeds is not None) and (inputs_embeds.shape[0] > 1)
         ):
-            raise ValueError("Generation with batch size > 1 is not currently supported!")
+            raise ValueError(
+                "Generation with batch size > 1 is not currently supported!"
+            )
 
         # Handle `past_key_values` (cache) =>> assume `input_ids` just has unprocessed tokens
         if past_key_values is not None:
@@ -789,7 +610,6 @@ class OpenVLAForActionPrediction(PreTrainedModel):
         )
 
         return model_inputs
-    
 
 
 EntryClass = OpenVLAForActionPrediction
